+++ src/Request/Checkout/Amount/Amount.php
@@ -27,7 +27,7 @@ class Amount extends Base
     /**
      * @inheritDoc
      */
-    public function jsonSerialize()
+    public function jsonSerialize(): mixed
     {
         return [
             'value' => $this->value,