+++ src/Request/Checkout/Checkout.php
@@ -89,7 +89,7 @@ class Checkout extends Base
     /**
      * @inheritDoc
      */
-    public function jsonSerialize()
+    public function jsonSerialize(): mixed
     {
         return [
             'id' => $this->id,