+++ src/Request/Checkout/Item.php
@@ -37,7 +37,7 @@ class Item extends Base
     /**
      * @inheritDoc
      */
-    public function jsonSerialize()
+    public function jsonSerialize(): mixed
     {
         return [
             'name' => $this->name,