+++ src/Request/Checkout/TotalAmount.php
@@ -33,7 +33,7 @@ class TotalAmount extends Base
     /**
      * @inheritDoc
      */
-    public function jsonSerialize()
+    public function jsonSerialize(): mixed
     {
         return [
             'value' => $this->value,