+++ src/Request/Webhook/Webhook.php
@@ -44,7 +44,7 @@ class Webhook extends Base
     /**
      * @inheritDoc
      */
-    public function jsonSerialize()
+    public function jsonSerialize(): mixed
     {
         return [
             'id' => $this->id,